Moving, renaming queue.
Moves the queue code into the `state` directory where it belongs.
