+++ imls-raspberry-pi/cmd/session-counter/state/queue_test._go
@@ -1,7 +1,7 @@
 package state
 
 import (
-	"testing"
+	"os"
 
 	"github.com/stretchr/testify/suite"
 )
@@ -11,38 +11,35 @@ type QueueSuite struct {
 }
 
 func (suite *QueueSuite) SetupTest() {
+	temp, err := os.CreateTemp("", "queue-test.ini")
+	if err != nil {
+		suite.Fail(err.Error())
+	}
+	SetConfigAtPath(temp.Name())
 }
 
 func (suite *QueueSuite) AfterTest(suiteName, testName string) {
+	os.Remove(GetDurationsPath())
 }
 
 func (suite *QueueSuite) TestQueueCreate() {
-	NewQueue[string]("queue1")
+	NewQueue("queue1")
 }
 
 func (suite *QueueSuite) TestEnqueue() {
-	q := NewQueue[string]("queue1")
+	q := NewQueue("queue1")
 	q.Enqueue("123")
 	q.Enqueue("abc")
 }
 
 func (suite *QueueSuite) TestMultiEnqueue() {
-	q := NewQueue[string]("queue1")
+	q := NewQueue("queue1")
 	q.Enqueue("123")
 	q.Enqueue("123")
-}
-
-func (suite *QueueSuite) TestQueueLength() {
-	q := NewQueue[string]("queue1")
-	q.Enqueue("123")
-	q.Enqueue("123")
-	if q.Length() != 2 {
-		suite.Fail("queue is the wrong length")
-	}
 }
 
 func (suite *QueueSuite) TestPeek() {
-	q := NewQueue[string]("newqueue")
+	q := NewQueue("newqueue")
 	_, err := q.Peek()
 	if err == nil {
 		suite.Fail("peek on an empty returned nil")
@@ -50,8 +47,7 @@ func (suite *QueueSuite) TestPeek() {
 }
 
 func (suite *QueueSuite) TestDequeue() {
-	q := NewQueue[string]("queue1")
-	q.Enqueue("abc")
+	q := NewQueue("queue1")
 	shouldremove, _ := q.Peek()
 	removed, err := q.Dequeue()
 	if err != nil {
@@ -60,8 +56,4 @@ func (suite *QueueSuite) TestDequeue() {
 	if removed != shouldremove {
 		suite.Fail("did not find appropriate next item.")
 	}
-}
-
-func TestQueueSuite(t *testing.T) {
-	suite.Run(t, new(QueueSuite))
 }
