+++ imls-raspberry-pi/cmd/session-counter/state/queue_test.go
@@ -1,7 +1,7 @@
 package state
 
 import (
-	"os"
+	"testing"
 
 	"github.com/stretchr/testify/suite"
 )
@@ -11,35 +11,38 @@ type QueueSuite struct {
 }
 
 func (suite *QueueSuite) SetupTest() {
-	temp, err := os.CreateTemp("", "queue-test.ini")
-	if err != nil {
-		suite.Fail(err.Error())
-	}
-	SetConfigAtPath(temp.Name())
 }
 
 func (suite *QueueSuite) AfterTest(suiteName, testName string) {
-	os.Remove(GetDurationsPath())
 }
 
 func (suite *QueueSuite) TestQueueCreate() {
-	NewQueue("queue1")
+	NewQueue[string]("queue1")
 }
 
 func (suite *QueueSuite) TestEnqueue() {
-	q := NewQueue("queue1")
+	q := NewQueue[string]("queue1")
 	q.Enqueue("123")
 	q.Enqueue("abc")
 }
 
 func (suite *QueueSuite) TestMultiEnqueue() {
-	q := NewQueue("queue1")
+	q := NewQueue[string]("queue1")
 	q.Enqueue("123")
 	q.Enqueue("123")
 }
 
+func (suite *QueueSuite) TestQueueLength() {
+	q := NewQueue[string]("queue1")
+	q.Enqueue("123")
+	q.Enqueue("123")
+	if q.Length() != 2 {
+		suite.Fail("queue is the wrong length")
+	}
+}
+
 func (suite *QueueSuite) TestPeek() {
-	q := NewQueue("newqueue")
+	q := NewQueue[string]("newqueue")
 	_, err := q.Peek()
 	if err == nil {
 		suite.Fail("peek on an empty returned nil")
@@ -47,7 +50,8 @@ func (suite *QueueSuite) TestPeek() {
 }
 
 func (suite *QueueSuite) TestDequeue() {
-	q := NewQueue("queue1")
+	q := NewQueue[string]("queue1")
+	q.Enqueue("abc")
 	shouldremove, _ := q.Peek()
 	removed, err := q.Dequeue()
 	if err != nil {
@@ -56,4 +60,8 @@ func (suite *QueueSuite) TestDequeue() {
 	if removed != shouldremove {
 		suite.Fail("did not find appropriate next item.")
 	}
+}
+
+func TestQueueSuite(t *testing.T) {
+	suite.Run(t, new(QueueSuite))
 }
